+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 
 ### FEATURES:
+* builder/AmazonInstance: add support for packer's amazon instance builder
 * builder/ProfitBricks: add support for packer's ProfitBricks builder
 * builder/Qemu: add support for packer's qemu builder
 * builder/Triton: add support for packer's triton builder